move 5.7.1 migrations to 5.8
